+++ train.py
@@ -10,7 +10,7 @@ from torch.utils.data import DataLoader
 from transformers import RobertaTokenizerFast
 from transformers.data.data_collator import default_data_collator
 
-from datasets.reuters_text import R8, R52
+from data_prep.reuters_text import R8, R52
 from models.model import ClassifierModule
 
 # disable parallelism for hugging face to avoid deadlocks
